requirements.txt
openai==1.68.2
python-dotenv==1.0.1
pydantic==2.10.6
hdbscan==0.8.40
nltk==3.9.1
numpy==2.2.4
pandas==2.2.3
scikit-learn==1.6.1
beautifulsoup4==4.13.3
requests==2.32.3
newspaper3k==0.2.8
sentence-transformers==3.4.1
lxml_html_clean==0.4.1
readability-lxml==0.8.1
pyspellchecker==0.8.2

google-api-core==2.24.2
google-auth==2.38.0
google-cloud-bigquery==3.30.0
google-cloud-core==2.4.3
google-crc32c==1.7.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
